+++ scripts/git-subtree/monaco-editor/add.ps1
@@ -1,0 +1,5 @@
+git subtree add `
+    -P workspace/plugins/monaco-editor `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-monaco-editor.git `
+    dev `
+    --squash