+++ scripts/git-subtree/monaco-editor/push.ps1
@@ -1,0 +1,4 @@
+git subtree push `
+    -P workspace/plugins/monaco-editor `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-monaco-editor.git `
+    dev