+++ scripts/git-subtree/monaco-editor/split.ps1
@@ -1,0 +1,3 @@
+git subtree split `
+    -P workspace/plugins/monaco-editor `
+    --rejoin